python-backend-starter/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
pydantic==2.5.3
numpy==1.26.3
opencv-python-headless==4.9.0.80
python-dotenv==1.0.0
PyJWT==2.8.0